Split iOS Mac picker test helpers
