+++ Packages/iOS/CmuxMobileShellUI/Tests/CmuxMobileShellUITests/WorkspaceMacSelectionIdentityProvider.swift
@@ -1,0 +1,10 @@
+import CmuxMobileShellModel
+
+@MainActor
+final class WorkspaceMacSelectionIdentityProvider: MobileIdentityProviding {
+    var currentUserID: String?
+
+    init(userID: String?) {
+        self.currentUserID = userID
+    }
+}
